Move limit_param_set from VPFuzzer to ParamRunner
