+++ custom_components/test/fuzzing/fuzzer_utils/ValuePoolFuzzer.py
@@ -1,7 +1,5 @@
 from itertools import product
 import logging
-import random
-
 
 from custom_components.test.fuzzing.fuzzer_utils.Fuzzer import Fuzzer
 from custom_components.test.fuzzing.fuzzer_utils.ValuePool import ValuePool
@@ -80,34 +78,3 @@ class ValuePoolFuzzer(Fuzzer):
 
         # print(result)
         return result
-    
-    def limit_param_set(self, param_set: list, runs: int) -> list:
-        """Generates a specific selection of an individual value pool. A list of lists is returned with a specified number of elements.
-
-        :param param_set: The value pool as list of lists.
-        :type param_nr: list
-        :param runs: Number of elements selected from the value pool.
-        :type types: int
-
-        :return: A random selection of certain elements from the parameter set.
-        :rtype: list
-        """
-
-        # Validate input parameters
-        if not isinstance(param_set, list):
-            self.logger.error("Param_set must be of type list.")
-            raise TypeError("Param_set must be of type list.")
-        if len(param_set) == 0:
-            self.logger.error("Length of param_set must be greater then 0.")
-            raise ValueError("Length of param_set must be greater then 0.")
-        if not isinstance(runs, int) or runs <= 0:
-            self.logger.error("Runs must be of type int and greater than 0. Parameter set is returned unchanged.")
-            return param_set
-        
-        # Selection of random elements from param_set if the number of runs is smaller than the number of elements in param_set
-        if runs > len(param_set):
-            self.logger.info("Length of param_set is smaller than the value of runs. Returned param_set unchanged.")
-            return param_set
-        else:
-            self.logger.info(f"Decresed elements in param_set from {len(param_set)} to {runs}")
-            return random.sample(param_set, runs)
